requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
Deprecated==1.2.14
exceptiongroup==1.2.2
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
future==1.0.0
h11==0.14.0
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
line-bot-sdk==3.11.0
MarkupSafe==2.1.5
matplotlib==3.9.1
multidict==6.0.5
numpy==2.0.1
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
requests==2.31.0
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
